Added error detection and unsigned signed stuff

+++ ECE574_Assignment2/src/TopModule.h
@@ -36,7 +36,7 @@ class TopModule {
 		void printModuleName(ofstream& circuitFile, string CircuitName);
 		void printModules(ofstream& circuitFile);
 
-		float findInputDelay(Module module);
+		double findInputDelay(Module module);
 		void findCriticalPath();
 
 };
